Fix: chatID column should be bigint
Apparently some chat ids coming from the Telegram API are too big to fit in a regular integer.

+++ database/migrations/2017_03_21_020000_create_table_Stalks.php
@@ -15,7 +15,7 @@ class CreateTableStalks extends Migration
     {
         Schema::create('Stalks', function (Blueprint $table) {
             $table->increments('stalkID');
-            $table->integer('chat');
+            $table->bigInteger('chat');
             $table->integer('uvaID');
             $table->timestamp('createdAt')->useCurrent();
             $table->timestamp('deletedAt')->nullable();
